Setup client/server integration:
React will be used as for the client, but it will still be rendered
by Django. Added integration to settings.py to make Django use the
generated deployable code as the template and added generated static
content to Django's static middleware, using whitenoise.

Added a Login template for authenticated pages to redirect to, if the
user isn't already logged in.

NOTE: The redirection isn't working, It always just goes to the client.