feat(frontend): Rewrite backend
The previous backend has some generic stuff baked in e.g. users, this
is something that is now a 3rd party app and should be imported like
any other dependency.

This is very basic implementation, we have user management, but I've
still to rewrite the client-side code and tie the two together.